requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
et-xmlfile==1.1.0
flake8==7.0.0
flake8-isort==6.1.1
fonttools==4.50.0
idna==3.6
isort==5.13.2
kiwisolver==1.4.5
matplotlib==3.8.4
mccabe==0.7.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pycodestyle==2.11.1
pyflakes==3.2.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-gitlab==4.4.0
pytz==2024.1
requests==2.31.0
requests-toolbelt==1.0.0
seaborn==0.13.2
six==1.16.0
tzdata==2024.1
urllib3==2.2.1
